add new diagrams and update docs
I added the new diagrams and updated the docs to include them along with
an explanation of what's going on for the simpler overview.

I included the graphml files so they can be imported and edited/improve
later on but I don't know how they'll work in other programs; I created
them in yEd: yworks.com/products/yed

The dark diagrams have transparent backgrounds so they should be
suitable for any shade and the light diagrams are simply white. SVGs are
including for editing colors and exporting at various sizes.