integrate vale
This massive patch attemps to standardise out documentation. For those
unfamiliar with it, Vale is "a syntax-aware linter for prose built with
speed and extensibility in mind" [1]. I built on Vale's
implementation [2] of Google's style guide [3] with specifics for Arbor
and the vocabulary we've started using.

[1]: https://github.com/errata-ai/vale
[2]: https://github.com/errata-ai/Google
[3]: https://developers.google.com/style/

I didn't modify any of the docs that are already written so some of them
will have a ton of errors. Once Arbor's style has been reviewed and a
bit more polished, I'll go through and start implementing the
recommended changes. To that end, install Vale [1], run it on a few
files, and see what the output says. Feedback is welcome.

[1]: https://errata-ai.gitbook.io/vale/getting-started/installation